climate-energy-analytics/docker/api/requirements.txt
fastapi
uvicorn
requests
boto3
kafka-python